refactor: renamed bd
